+++ src/metric.c
@@ -2,7 +2,7 @@
 #include "metric.h"
 #include "debug.h"
 
-void points_add(struct points *points, short x, short y) {
+void points_add(struct points *points, int x, int y) {
     NULL_CHECK("points_add()", points);
 }
